Renamed the controller tests for GetProcessData method.
Did this so that the names of these tests would keep on making sense after more controller methods (endpoints) are introduced into that same controller.

This change will most likely cause a merge conflict due to the base brach of this branch not being merged to master yet (due to pending reviews).