Added a test that covers the temporary endpoint implementation's Response.
It was decided that for the time being only the controller method will be implemented, that would pretend to have done the work that it actually should do. Essentially a it would return a success response, which in case of POST is "201 Created".

Reason for it was that this was needed for the front-end part of the project to progress. (Have something to talk to).

Absolutely no implementation yet, so this commit should have this test as failing test.