Added a PostInitialProcessDocument controller method. Made it always return 201 status code.
Implemented just a success response as requested.

While doing this I noticed that our Swagger documentation does not cover the Response for successful creation.

Typically you would expect to get your posted object back, as well as a URL, where it was created at. But I'm leaving this up for discussion and putting some dummy data in, as some sort of object in the response is required anyway.