Added a test that covers temporary UpdateExistingProcessDocument endpoint implementation.
Same as with the PostInitialProcessDocument endpoint.

The idea is to pre-implement an endpoint to return success results for the front-end to have something to talk to before the actual implementation is done.
So that this wouldn't slow the front-end work too much.

Since this is only a test and no implementation, the tests will fail.