Changed endpoint to Patch method and return 200 on Success.
Changed the test to reflect that, as well as the implementation.

Also changed the endpoint to PATCH http method because according to our guidelines, PUT shouldn't be used that way.

There was no discussion about the status codes yet (barely anyone is in), but at least for the time being, or maybe permanently it was decided to set it to "200 Ok".