Added a method overload for creating MatProcessData Domain object
-----Implementation------
Had to extract dateOfCreation out, since I believe that the dates DateCreated and DateLastModified have to be equal. The alternative is them differing by miliseconds, which might or might not cause issues in the future - who knows how it's going to be used.

The process stage... I left it as a string because I was told that it's still undecided on its format. Therefore I don't feel like changing a lot of existing code around that for an uncertainty.

DateCompleted - a minimum value. Not elegant, but it was decided that it's better than having a nullable DateTime type.

Why not an instance method? Was following the previous method in the factory. Don't understand enough C# to question that decision yet. I guess, it's all good as long as this method doesn't hold any state or anything.

------Tests--------
Using reflections because the regular comparison that's commented out returns a nonsensical error saying that it expected <{ }>, but instead got <{ }>.

It might have been simply comparing object references, but who knows.

So instead I check how many fields the anonymous object has. It's supposed to be zero, since the requirement is that it has to be set as empty object.