Added a PostInitialProcessDocument endpoint Usecase.
--------- Tests ----------
Did not include all of the domain object property fields in the mock gateway verify assertion because test doing that is in the Factory tests. Instead I aimed to show that the object passed into the gateway contains some fields from the request object, and some fieds generated from the factory... to show that request object data was mapped onto domain object, as well as to show that it was the factory that did it by including a field that only it generates (that's not a default of some type in C# (like 'int' default is 0)).

This usecase has only a single test, since it's only invoking other methods and not expecting or returning any output.
The existing test could have been split into 3 tests if the we hadn't agreed on making factory methods static. The issue is that static methods can not be mocked, therefore you can't check if there was a call made to a static method.
The way the test could have been split would have been: doing a test to see if usecase calls the factory.
The other one to see if use case calls the mock gateway with the mock factory output.
The 3rd test would be similar to 2nd one except the factory is not a mock any more - this would add more of a real world element to it... it would still call mock gateway in the end.

----- Gateway -----
A blank method had to be included for the project to compile for a usecase test to run. Same method signature had to be included in the interface.

---- Usecase -----
Usecase has 2 things to do for this endpoint:
1) Call the thing that maps the data
2) Use that data to call the gateway that will insert it.
Reason for usecase calling the mapping the data is because we're using MadeTech Flavoured Clean Architecture. As per that architecture the Gateway should only be receiving the Domain object, not the request object. Therefore it was decided to put the call to mapping into the usecase.