Added Gateway tests for POST initial process document endpoint
There should be an extra tests, as well as some changes made to the existing tests due to the recent change in specification.

Not doing all these things now, because doing the changes to match the specification will break existing tests, and I'd rather keep that to a separate commit for the sake of cleaner commits.