PostInitialDocument gateway method basic implementation
Did just enough to get the tests pass. Tests only covered the insertion of the document, and not the try catch error handling, so there still is some work to do.

In the Second test for checking if it was just a single document inserted, there is this variable unclearedDocumentCount. It's there because it looks like the database doesn't get cleared after every test, so it makes it so that whatever records were inserted by other tests prior to this test, get to stay and ruin the day.
So you need to capture the initial document count.
Then I generate random amount of other documents for triangulation.
And only then I insert my actual document I'd want to insert IRL, and check if the count is 1 greater, than documents pre-inserted before test + documents pre-inserted before the main test's insertion.

The gateway method is no longer void since the specification changed from it giving no response to it giving just the processRef (Id)