Made Post Usecase wrap up and return gateway output.
Like last time, one of the tests is impossible due to decision of sticking to 'static' factory rather than instance.

Added additional to test to cover the case of the usecase returning something. According to the new requirements, this endpoint is supposed to return the processRef (Identifier for the document) rather than empty response. Essentially 201 over 204. Inside the test had to resort to some trickery for returning something as simple as a string due mock acting in weird ways.

Changed few types in the PostInitialProcessDocumentResponse object, which was  unnecessarily added last time as it's only now that it's useful.

Implemented the return and wrapping logic in the usecase implementation file.