Changed ">" to "|" to avoid issues when copy-paste'ing pipeline
Apparently ">" combines anything together that's separated with only 1 newline (no empty line in between)
so you have to add empty lines in between the comments and commands for them not to be appended to the 1st line which is the comment.

Use  the "|" which will interpret every newline as a different command. To use this I also had to add the "\" to that big command between the comments, otherwise it would be interpreted as 2 commands and crash, rather than 1 command across 2 lines. This is also better in the case where copying, pasting removes any successive newlines after the 1st one, which would trip the whole command when ">" is used.