TDD'ed basic PostInitialProcessDocument Controller method implementation
No error handling or validation yet - the format is still to be decided.

To proceed further with controller need to implement the validator class as well as some basic error handling and wrapping.

This commit contains the acceptance and unit tests for success cases as well as comments as a reminders for what other tests need to be in there once all other necessary pieces get implemented.

Change the controller return to ObjectResult. Reason for it is: while we decided that we want the 201 status code, the specification wants to return only one property of the object inserted rather than url to it + the object itself.

In processDataGatewayTests I removed the dev line left over from debugging.

In processDataGateway added a reminder to change "InsertOneAsync" to "InsertOne" as it was noticed that the async method does not detect duplicate inserts.

All test pass on my machine, however I don't have a way to test after having launched application up - due to missing mongo connection string??