Fixed bug - empty {error message} part of the exception string.
A test reveals a bug, where the InnerException has not been passed into custom exception, hence information loss, and incomplete custom error message.

The fix, create a constructor for custom exception that takes inner exception as an argument. Which allows to pass the previously missing information in.

The custom exception should probably be extracted away into some exceptions folder, but that's probably not for now.