Removed unneeded [Required] annotations. Renamed variable for more clarity. Removed some typos, unneeded comments. Added some needed comments.
Changed 'toBeInsertedDocumentCount' into preInsertedDocumentCount, so that it wouldn't be confused with 'toBeInsertedDomainObject', which is the start of the test.

Reasoning behind the [Required] removal is that ... if  for some reason any of these properties would not need to be Required anymore... you'd have to change it in the validations (where you expect) and in the here (where you wouldn't) - it's 2 places doing same thing - not very good anyway.
Also if anyone were to need to add another property, this would cause confusion due to there being 2 places that do this, yet this one being inactive.
Tests still pass as they should.