Fixed few issues regarding MatProcessData generation in the Helper method. Also Added few comments in the Validator and Gateway explaining few less clear parts of the code.
----------------- Validator -----------------
Fixed the duplicate Rule for 'ProcessType.value'. It had '.NotEmpty()' twice, so I changed one of them into '.NotNull', Which might become useful if integer becomes nullable, or gets set to null in the request somehow.

Added the comment explaining the CascadeMode validator option.

----------------- Gateway -----------------
Added a comment explaining the AggregateException handling.

----------------- Helper -----------------
The 'ProcessDataSchemaVersion' random value between (0,10) combined with the not zero test in the validation itself, has a probability of 10% of failing any controller test that uses this helper. There aren't any yet, but as a good precaution, it's best to do changes before issue happens.
The starting Schema version will be '1', so need to exclude zero anyway, if not for the issue stated, then for consistency.

The way MatProcessData object has been used does not matter in the context of ProcessStage. All that's being checked is whether the inserted docs data matches the retrieved one's. Therefore I can change this helper method to generate the "Not Completed" value for ProcessStage rather than the random value. It's better this way for consistency reasons, because the method in the edit is the method for generating **initial** document as can be seen from from other properties it generates.