Completed the incomplete Gateway test that checks whether the Usecase calls the gateway with the correct data.
This test is supposed to check whether the Usecase calls the Gateway with whatever ProcessDataFactory converter (from Request obj to Domain obj) would output.

However, previously the test was only checking whether the Usecase calls the Gateway with the correct type of the object. This excluded the check whether the data was correct.

So in this commit, I add a check for the Data contained in the object as well:

In the Verify bit you can check whether the mock object is called with the correct properties of the passed in object. Unfortunately checking the object itself (which would be a shortcut to typing every property check) does not work due to C# checking whether object references match rather than whether object contents match:
The call to factory in the UseCase gives ref1, while the call in the test (domainObject) gives ref2, hence test fails, if you don't check the contents of the properties themselves instead.