Added Post Initial Document Request Validator into dependency injection setup. And did the compatability changes.
Create an interface for the PostInitialDocumentRequestValidator.

Made that validator an instance of that interface. Made an interface require '.Validate()' method, as it's the thing that's being used from within the controller.

In the controller I added validator as a parameter, rather than having it set to instance in the constructor.

In the tests, in order to match new definition of controller, I added mocks to deal with that. Where needed I set mocks to return a no-validation-error result, because all the tests at the time were of success cases only.