Made the skeleton Post Process Image Controller.
Paired up with Ana Bebic - was showing the ways of how API is structured.

We were working on creating an endpoint to store images and do something else that's still not exactly clear due to the planing session not being done yet.

Added the Controller of the POST image endpoint as well as a Boundary object and Usecase interface, which was necessary to write the tests.
The controller is mostly done, only need to add a call to validation and some error handling.

It was decided that, at least, for the time being the endpoint will not be returning anything, hence the NoContent 204.

Reasoning is that the image data will passed around using base64 string, which we expect to be massive. Having a response which returns the request (like typical) is out of the question, since encoded image string is sort of useless and would take quite some network resources to pass around. Same reason we don't return the inserted image itself. So there's nothing to return.

That's not exactly true, there still is an option to return a link to the image and maybe the imageId based on the implementation of the gateway we decide.

The path of the new is still undecided, for the time being temporary one is set.