Made a Usecase for Persist Image endpoint
Had to create a 'ProcessImageData' domain object, because we don't want the gateway to receive a base64 image string. It is not its' responsibility to do the decoding.

Instead we have a domain object that includes already decoded base64 string into byte[], which we found out from some research can be used to insert object into S3.

https://stackoverflow.com/questions/48634807/amazon-s3-net-upload-base64-image-data

We also had to include (image) decoder call in the usecase as well as make decoder's interface so that we would be able to proceed with the usecase test of it calling the gateway with correct data (and data type).