Added MatProcessDataHelper method to generate a PostProcessImageRequest
Added to improve readability and to make it easier to maintain.

Change the arrange in the tests accordingly.  Left an empty request newing up within the tests because it illustrates that test is not dependent on the contents of the request, also it easier to read to due shorter code line.