Added Boundary Response object for Get Process Image endpoint
As discussed, follows typical pattern of giving back the request and generatedAt time in addition to an actual data being returned.

In this case the data returned is simply a string due to the simplicity of endpoint. This means there will probably be no need for the Domain object to wrap this up.