Added 'Unexpected' Error Handling logic to ProcessImageController.
Wrote a test that checks whether the information form an error was put and structure and wrapped up into response the way we expect. Test generates 4 possible errors triangulation.

Added implementation. For now it's only for 500 responses. I'd imagine there will be more specific exceptions to catch in the future, however practice show they typically originate from the gateway, hence it's for later.