Factored out error throwing into ErrorThrowerHelper and improved it.
- Created a new tests helper for throwing errors so that error handling would be easier to test.
- Improved the ErrorThrowerHelper to include errors with not null Inner Exceptions.
- Improved the ErrorThrowerHelper to require randomised string for error message triangulation
- Improved both unexpected error handling tests to make use of this helper.

Reason why I don't have random value generation within the helper is because during tests I typically need to call it twice, and I need it so both calls would give the same error message, which was random. The only way to do that is to randomize the message prior to the call and drop it in as an argument.