Added Logger to the 'Process Image Controller' for Get Endpoint
- Logger logs that the request happened.
- Logger logs that the request didn't pass due to validation failure, and returns a formatted message of the errors returned.

- Testing Logger was a bit tricky due to Logger.LogInformation method (and others) being 'extension methods' - whatever that would mean. The thing is, they can't be mocked, so Moq fails both on the .Setup() as well as .Verify(), so what's left is:

- The way to mock an extension method is not to mock an exte nsion method, you instead mock the real method. In this case .Log() with it's biggest constructor. So what actually ends up happening is, instead of checking if the extension method '.LogInformation()' got called, you instead check if the extenstion method called the actual '.Log(LogLevel.Information, ..., ...,.....)' method.

The logic is - if the extension method is called with said parameters, then the real method will be called by that extension method with its parameters passed in, in addition to other hidden parameters that extension method had.

There's slightly better way to do this as I've seen online, however due to time restraints, I won't be able to test those out due it actually being a spike and requiring unknown amount of time.

'https://ardalis.com/testing-logging-in-aspnet-core'