Fixed the PostProcessImageRequestValidator Test from sometimes failing - Modified the request generator.
The test was that there should be no error when the base64 string of correct format was provided. The base64 string provided was generated by this method here in the tests helper.

The issue was that faker file extension generator was sometimes generating file extensions of shorter length than allowed. Due to the method not being flexible enough, had to change it into random string generator.

Had to use string2 and provide char array of allowed characters due to the 'string' method being hard to use for alphanumerics (would have to include @ symbol and something else, if I specified the range from '\u0031' to '\u005a'. Which is not allowed character for a file extension.