Improved tests readability by explicitly stating that test does not depend on parameters of request object
The reason why this line:

`var request = new GetProcessImageRequest();`

was left as is rather than changed to be generated from helper was that it helps to signify that this test does not depend on what the request object is, and that the only reason the request object is there is to keep the compiler happy so that the test could be run.

Clarification: even though some of these tests changed here have `given_a_valid/invalid_request_object_...` in their name, it does not mean that the test itself depends on that object. In most these cases change here, test actually depends on the mock of some dependancy that would would give some sort of response under the conditions of `given_valid/invalid_requets_...`.

The reason why test has that bit in it's name is to justify the mock setup (and what it returns), as that mock under real circumstances should be behaving in a that way when the controller is given valid/invalid response.

If we were to make the tests in a way that the contents of the request object were of some significance, we would fall into the trap of seting up a chain of mocks that have no choice but to have actual implementation within them. Under these circumstances, the test would no longer be testing just that one thing it says it tests, but it would be testing almost the entire controller end to end, which is not what we want from unit tests.

Reasoning behind these particular changes:
- Ln192: "Test only depends the validator mock setup"
- Ln304: "Test about catching an exception, which we throw manually"
- Ln341: "Test about catching exception messages that we generate from exceptions that we manually throw."
- Ln360: "Test about calling the usecase if validator permits it - test depends only on validator mock setup."
- Ln406: "Test about calling the logger unconditionally at least once. (like when any request comes in - even if it's empty)"

Writing this all out made me think that I should improve the test names to better reflect what they're testing.