Modified the GetProcessImageRequest Validator.
Added 2 extra tests for the 2 new fields of GetProcessImageRequest object ("processType, fileExtension). These tests are about throwing an error when a blank parameter is supplied. Added implementaion to validator based of these tests.

Also added back the removed tests and validation for the "processRef" and "imageId" fields. The ones that did the not null, not empty checks. Reason for it is that even though request with such null values wouldn't pass the validation, it wouldn't give the error message specific enough to what the issue is. I feel that providing 'null' argument versus an invalid one should return different error messages.

Now, there was an argument raised on the PR against these checks saying that: since now the parameters are being passed in through url path, when you don't provide them, it will simply be interpreted as a different path - hence 404 path not found.
However, in response to that, I'd say: if you had provided a url encoded whitespace (which sometimes gets encoded automatically) as an argument, it would count as a valid path, and the  .NET would interpret the whitespace as 'null'. I checked it through the browser and debugger - it confirms it.

Example: "{insert your host}/api/v1/processImageData/%20/a68e234a-bd08-3542-4b89-42572b6adc65/e68e234a-bd08-3542-4b89-42572b6adc65/%20"