Renamed the GetProcessImage endpoint Controller tests to be more desciptive about what they actually test.
-------------- Request unimportant --------------

In a lot of cases where the test names starts with: "given_a_valid_request" or "given_an_invalid_request". The request object created or referenced in the name does not matter at all. Example:
Ln189 "given_a_valid_request_when_GetProcessImage_controller_method_is_called_then_it_returns_a_200_Ok_response"
this test name was misleading because due to this being a unit test - all the dependencies of class under test (in this case controller) are being mocked. The mock itself contains no logic apart from whatever you make it return when doing mock setup in the arrange step. So in this test's case it means that it does not matter what request I pass into it, it will always return whatever I specified that mock to return. So, if I'm testing "given_a_valid_request", what it actually means is that I'm testing whether whatever is specified happens based on request validation logic returning success. So the test is more like "under_the_validation_control_flow_of_success_X_happens".

To make this explicit, I renamed the test as well as moved empty object creation directly into argument place in the controller call (act phase). To signify that the request itself is not part of the test context, hence shouldn't be in the arrange step.

There are many more cases like this, so I won't cover them.

------------------Status Codes------------------

Ln263 status code wasn't even being used, because the test didn't even check for that - that was the job for the test before it. It's good to have these tests separated for easier and quicked troubleshooting.

Ln289 Since the test states that it's checking for 500 status code result in its name, this means that it should probably be put in the arrange step to make this more explicit.

--------------------"Correct Data"--------------------

I figured that it's a pretty bad description for whatever you're expecting as a result from the test. This is because this 'term' is too abstract and can mean too many things. So I changed this term into something more specific accross the Get endpoint tests in order to clarify what we're actually testing for.

----------------Modified few logger tests----------------

Tests in lines: "Ln423" -> "Ln411" and "Ln445" -> "Ln431" were indeed testing what they said, however they did not reflect the intention behind the tests.

The point of those tests was to show that logger is being called regardless of validation controll flow - as in it was called prior to the validation, however all they were testing was that the logger was being called at least once accross the entire validation controll flow.

So I changed these test to test exactly what they meant - that the request gets registered before it gets validated, meaning this registering action does not depend on the validation result.

To do this I had to ensure there's no Validator mock setup, which would influence the decision to call logger or not. This of course means a crash within the controller, since it's trying to make a decision based on null result now, but this crash can get mitigated by putting an empty try-catch block to ignore the error.

The whole idea is:
- If logger call depends on the validation result, due to a crash (and absense of control flow decision) in the controller, the call  to logger will never happen. And the test will fail due to there being 0 calls to logger in the assert step.
- However, if the logger call is independant of validation controll flow, then we should expect the call to happen prior to validation (prior to the error), and the test passes.

Now the counter argument could be: "what if the logger get called after all the validation, or in every branch of validation decision?". I'd say that even under these circumstances, you're still dependent on the validation - maybe not on the decision it makes, but on it not crashing, which is a bad way to ensure that every request coming into the controller gets registered.