Modified GetProcessImage logger tests and logger.
Modified the tests to expect 4 properties logged from the GetProcessImageRequest object. (Previously it was only 2, only after that additional 2 were added.)

Added the implementation to make modified tests pass.

Modified couple names of the logger tests again to be more explicit that this logger call does not depend not only on the result of the validation, but the behavior (like crashing) as well.

Added additional test that ensures that the initial logger call will also have null values from the request replaced with "null" strings. (In order to have some consistency accross the calls; Also helps with troubleshooting )